REF/FIX: reworked main -> now using PreVMain(...)
