+++ PreVEngineTest/platform/common/Main.cpp
@@ -1,0 +1,7 @@
+// forward client main declatation
+int PreVMain(int argc, char** argv);
+
+int main(int argc, char** argv)
+{
+    return PreVMain(argc, argv);
+}
